AXC6700: GPDMA: dts: indicate GPDMA operations are coherent
Add device-tree property 'dma-coherent' to GPDMA indicating its operations
are coherent on a system wise. GPDMA is an ACE-Lite (AXI Coherent
Extension-Lite) Master meaning is coherent with the system memory and all
the caches of RN-Fs (Fully Coherent Requesting Nodes) which are
the clusters with ARM CPUs and DSP CEVA CPUs.

Without this the generic DMA API try to either clean a DMA buffer if
allocated by the CPU or invalidate if filled in by the DMA at buffer
mapping and unmapping operations. See __dma_map_area() and
__dma_unmap_area() in kernel/arch/arm64/mm/cache.S for more details.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>