bus: arm-ccn: split arm-ccn interrupts into the core and PMU.
It happens out there are other drivers of CCN to CCN Perf, such as an EDAC L3
detecting and reporting of the HW errors. As the ARM CCN resources are common,
that is control registers but particularity an interrupt line (INTREQ),
a need arose to take the interrupt handling out and create a separate (core)
driver managing it.

ARM CCN signals errors (corrected and any) and PMU event counter overflow
through INTREQ. If a client driver of either exists the driver shall register
its interrupt service routine with the core. The core then dispatches
the interrupt to the client respectively. If neither of the drivers exist it
simply reports the error/s and the clears the interrupt line.

This particular commit splits ARM CCN Perf to the core and PMU.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>