bus: arm-ccn: L3 EDAC: add another client driver of arm-ccn.
Update L3 EDAC driver into using the arm-ccn core. These changes
rely on the split of arm ccn Perf into the interrupt core and
the clients.

Along put a few fixes and enhancements to the EDAC code:
- Add struct device facilitating using the dev_<level> macros for
  printing out
- Add driver data to the device so that it can be obtained in
  the remove function (platform_{set,get}_drvdata)
- Change the compatible string from "intel,..." to "arm,...". DTSpec
  recommended format is "manufacturer,model", where manufacturer is
  a string describing the name of the manufacturer (such as a stock
  ticker symbol), and a model specifies the model number. In the end
  arm ccn is ARM's not Intel's IP.
- platform_driver's remove method mustn't un-register the device. It
  shall be left to the platform device core to do so, not to
  the driver. It should though release the resources allocated
  in the probe method.

Note also that now with the L3 EDAC using the interrupt core the error
interrupts will be process faster. Before the top half was hard irq,
bottom - the irq thread with IRQF_ONESHOT masking the interrupt until
the thread completes. Now all is handled in the hard irq.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>