bus: arm-ccn: L3 EDAC: add another client driver of arm-ccn.
Update L3 EDAC driver to the arm-ccn core.

These changes rely on the split of arm ccn Perf into the interrupt core and
the clients. Along put a few fixes and improvements to the EDAC code:
- Add struct device facilitating using the dev_<level> macros for printing out
- Add driver data to the driver_data of device so that the custom driver data
  can be obtained in the remove function
- Change the compatible string from "intel,..." to "arm,...". DTSpec recommends
  a format is "manufacturer,model", where manufacturer is a string describing
  the name of the manufacturer (such as a stock ticker symbol), and model
  specifies the model number. In the end arm ccn is ARM's not Intel's IP.
- platform_driver's remove method mustn't un-register the device. It shall be
  left to the platform device core to do so. It should though release the
  resources allocated in the probe method.

Note also that now with L3 EDAC using the interrupt core the error interrupts
will be process faster as in hard irq. Before the bottom part was being processed
in the irq thread with IRQF_ONESHOT masking the interrupt until the thread
completes.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>