axxia: GPDMA: Run GPDMA WorkAround Enabling the CCN Secure Access at EL2
We have a WA which enables Non-Secure access to the Secure Registers of
CCN via GPDMA. However, it has been shown it fails on some of the boards
in which it was successfully run at EL2 (Exception Level) from inside
U-Boot (bl33 by ARM notion). This patch implements the WA at EL2
scheduled from the Linux kernel but still first trying it out at EL1,
and only if unsuccessful, at EL2.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>